First shot at drawing sprites
We can draw sprites! There are a few problems - we need to respect
properties like horizontal mirroring or mario will look weird, and
there's some sort of vertical alignment bug. There's a slight fps drop
from this. Most of it comes from iterating through all the sprites on
the scanline for every pixel, and a bit of it comes from disabling the
background caching we had earlier.